Moved datasources to datasource dir, created shared FileBasedDatasource.
