+++ python/ray/experimental/data/datasource/datasource.py
@@ -1,10 +1,6 @@
 import builtins
 import logging
-from typing import Any, Generic, Optional, List, Tuple, Callable, Union, \
-    TypeVar, TYPE_CHECKING
-
-if TYPE_CHECKING:
-    import pyarrow
+from typing import Any, Generic, List, Callable, Union, TypeVar
 
 import numpy as np
 
@@ -17,98 +13,6 @@ T = TypeVar("T")
 WriteResult = Any
 
 logger = logging.getLogger(__name__)
-
-
-def _expand_directory(path: str,
-                      filesystem: "pyarrow.fs.FileSystem",
-                      exclude_prefixes: List[str] = [".", "_"]) -> List[str]:
-    """
-    Expand the provided directory path to a list of file paths.
-
-    Args:
-        path: The directory path to expand.
-        filesystem: The filesystem implementation that should be used for
-            reading these files.
-        exclude_prefixes: The file relative path prefixes that should be
-            excluded from the returned file set. Default excluded prefixes are
-            "." and "_".
-
-    Returns:
-        A list of file paths contained in the provided directory.
-    """
-    from pyarrow.fs import FileSelector
-    selector = FileSelector(path, recursive=True)
-    files = filesystem.get_file_info(selector)
-    base_path = selector.base_dir
-    filtered_paths = []
-    for file_ in files:
-        if not file_.is_file:
-            continue
-        file_path = file_.path
-        if not file_path.startswith(base_path):
-            continue
-        relative = file_path[len(base_path):]
-        if any(relative.startswith(prefix) for prefix in [".", "_"]):
-            continue
-        filtered_paths.append((file_path, file_))
-    # We sort the paths to guarantee a stable order.
-    return zip(*sorted(filtered_paths, key=lambda x: x[0]))
-
-
-def _resolve_paths_and_filesystem(
-        paths: Union[str, List[str]],
-        filesystem: "pyarrow.fs.FileSystem" = None
-) -> Tuple[List[str], "pyarrow.fs.FileSystem"]:
-    """
-    Resolves and normalizes all provided paths, infers a filesystem from the
-    paths and ensures that all paths use the same filesystem, and expands all
-    directory paths to the underlying file paths.
-
-    Args:
-        paths: A single file/directory path or a list of file/directory paths.
-            A list of paths can contain both files and directories.
-        filesystem: The filesystem implementation that should be used for
-            reading these files. If None, a filesystem will be inferred. If not
-            None, the provided filesystem will still be validated against all
-            filesystems inferred from the provided paths to ensure
-            compatibility.
-    """
-    from pyarrow.fs import FileType, _resolve_filesystem_and_path
-
-    if isinstance(paths, str):
-        paths = [paths]
-    elif (not isinstance(paths, list)
-          or any(not isinstance(p, str) for p in paths)):
-        raise ValueError(
-            "paths must be a path string or a list of path strings.")
-    elif len(paths) == 0:
-        raise ValueError("Must provide at least one path.")
-
-    resolved_paths = []
-    for path in paths:
-        resolved_filesystem, resolved_path = _resolve_filesystem_and_path(
-            path, filesystem)
-        if filesystem is None:
-            filesystem = resolved_filesystem
-        elif type(resolved_filesystem) != type(filesystem):
-            raise ValueError("All paths must use same filesystem.")
-        resolved_path = filesystem.normalize_path(resolved_path)
-        resolved_paths.append(resolved_path)
-
-    expanded_paths = []
-    file_infos = []
-    for path in resolved_paths:
-        file_info = filesystem.get_file_info(path)
-        if file_info.type == FileType.Directory:
-            paths, file_infos_ = _expand_directory(path, filesystem)
-            expanded_paths.extend(paths)
-            file_infos.extend(file_infos_)
-        elif file_info.type == FileType.File:
-            expanded_paths.append(path)
-            file_infos.append(file_info)
-        else:
-            raise FileNotFoundError(path)
-    return expanded_paths, file_infos, filesystem
 
 
 class Datasource(Generic[T]):
@@ -261,108 +165,6 @@ class RangeDatasource(Datasource[Union[ArrowRow, int]]):
                         schema=schema,
                         input_files=None)))
             i += block_size
-
-        return read_tasks
-
-
-class JSONDatasource(Datasource[Union[ArrowRow, int]]):
-    """JSON datasource, for reading JSON files.
-
-    Examples:
-        >>> source = JSONDatasource()
-        >>> ray.data.read_datasource(source, paths="/path/to/dir").take()
-        ... {"a": 1, "b": "foo"}
-    """
-
-    def prepare_read(self,
-                     parallelism: int,
-                     paths: Union[str, List[str]],
-                     filesystem: Optional["pyarrow.fs.FileSystem"] = None,
-                     **arrow_json_args) -> List[ReadTask]:
-        import pyarrow as pa
-        from pyarrow import json
-        import numpy as np
-
-        paths, file_infos, filesystem = _resolve_paths_and_filesystem(
-            paths, filesystem)
-        file_sizes = [file_info.size for file_info in file_infos]
-
-        def read_json(read_paths: List[str]):
-            logger.debug(f"Reading {len(read_paths)} files.")
-            tables = []
-            for read_path in read_paths:
-                with filesystem.open_input_file(read_path) as f:
-                    tables.append(
-                        json.read_json(
-                            f,
-                            read_options=json.ReadOptions(use_threads=False),
-                            **arrow_json_args))
-            block = ArrowBlock(pa.concat_tables(tables))
-            return block
-
-        read_tasks = [
-            ReadTask(
-                lambda read_paths=read_paths: read_json(read_paths),
-                BlockMetadata(
-                    num_rows=None,
-                    size_bytes=sum(file_sizes),
-                    schema=None,
-                    input_files=read_paths)) for read_paths, file_sizes in zip(
-                        np.array_split(paths, parallelism),
-                        np.array_split(file_sizes, parallelism))
-            if len(read_paths) > 0
-        ]
-
-        return read_tasks
-
-
-class CSVDatasource(Datasource[Union[ArrowRow, int]]):
-    """CSV datasource, for reading CSV files.
-
-    Examples:
-        >>> source = CSVDatasource()
-        >>> ray.data.read_datasource(source, paths="/path/to/dir").take()
-        ... {"a": 1, "b": "foo"}
-    """
-
-    def prepare_read(self,
-                     parallelism: int,
-                     paths: Union[str, List[str]],
-                     filesystem: Optional["pyarrow.fs.FileSystem"] = None,
-                     **arrow_csv_args) -> List[ReadTask]:
-        import pyarrow as pa
-        from pyarrow import csv
-        import numpy as np
-
-        paths, file_infos, filesystem = _resolve_paths_and_filesystem(
-            paths, filesystem)
-        file_sizes = [file_info.size for file_info in file_infos]
-
-        def read_csv(read_paths: List[str]):
-            logger.debug(f"Reading {len(read_paths)} files.")
-            tables = []
-            for read_path in read_paths:
-                with filesystem.open_input_file(read_path) as f:
-                    tables.append(
-                        csv.read_csv(
-                            f,
-                            read_options=csv.ReadOptions(use_threads=False),
-                            **arrow_csv_args))
-            block = ArrowBlock(pa.concat_tables(tables))
-            return block
-
-        read_tasks = [
-            ReadTask(
-                lambda read_paths=read_paths: read_csv(read_paths),
-                BlockMetadata(
-                    num_rows=None,
-                    size_bytes=sum(file_sizes),
-                    schema=None,
-                    input_files=read_paths)) for read_paths, file_sizes in zip(
-                        np.array_split(paths, parallelism),
-                        np.array_split(file_sizes, parallelism))
-            if len(read_paths) > 0
-        ]
 
         return read_tasks
